Apply some suggestions
