+++ docs/pages/x/api/charts/line-element.json
@@ -1,6 +1,5 @@
 {
   "props": {
-    "skipAnimation": { "type": { "name": "bool" }, "default": "false" },
     "slotProps": { "type": { "name": "object" }, "default": "{}" },
     "slots": {
       "type": { "name": "object" },
