+++ packages/x-charts/src/LineChart/LineElement.tsx
@@ -7,9 +7,10 @@ import generateUtilityClass from '@mui/utils/generateUtilityClass';
 import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
 import { SlotComponentPropsFromProps } from '@mui/x-internals/types';
 import { useInteractionItemProps } from '../hooks/useInteractionItemProps';
-import { AnimatedLine, AnimatedLineProps } from './AnimatedLine';
+import { AnimatedLine } from './AnimatedLine';
 import { SeriesId } from '../models/seriesType/common';
 import { useItemHighlighted } from '../hooks/useItemHighlighted';
+import { AnimatedLineProps } from '@mui/x-charts/hooks';
 
 export interface LineElementClasses {
   /** Styles applied to the root element. */
@@ -141,11 +142,6 @@ LineElement.propTypes = {
   d: PropTypes.string.isRequired,
   gradientId: PropTypes.string,
   id: PropTypes.oneOfType([PropTypes.number, PropTypes.string]).isRequired,
-  /**
-   * If `true`, animations are skipped.
-   * @default false
-   */
-  skipAnimation: PropTypes.bool,
   /**
    * The props used for each component slot.
    * @default {}
