+++ packages/x-charts/src/PieChart/PieArc.tsx
@@ -1,15 +1,12 @@
 'use client';
 import * as React from 'react';
 import PropTypes from 'prop-types';
-import { arc as d3Arc } from '@mui/x-charts-vendor/d3-shape';
 import composeClasses from '@mui/utils/composeClasses';
 import generateUtilityClass from '@mui/utils/generateUtilityClass';
 import { styled } from '@mui/material/styles';
 import generateUtilityClasses from '@mui/utils/generateUtilityClasses';
-import { interpolateNumber } from '@mui/x-charts-vendor/d3-interpolate';
-import useForkRef from '@mui/utils/useForkRef';
-import { ANIMATION_DURATION_MS, ANIMATION_TIMING_FUNCTION } from '../internals/animation';
-import { useAnimate } from '../internals/useAnimate';
+import { useAnimatePieArc } from '../hooks';
+import { ANIMATION_DURATION_MS, ANIMATION_TIMING_FUNCTION } from '../internals/animation/animation';
 import { useInteractionItemProps } from '../hooks/useInteractionItemProps';
 import { PieItemId } from '../models';
 
@@ -57,76 +54,6 @@ const useUtilityClasses = (ownerState: PieArcOwnerState) => {
 
   return composeClasses(slots, getPieArcUtilityClass, classes);
 };
-
-type PieArcAnimatedProps = Pick<
-  PieArcProps,
-  | 'startAngle'
-  | 'endAngle'
-  | 'cornerRadius'
-  | 'paddingAngle'
-  | 'innerRadius'
-  | 'outerRadius'
-  | 'skipAnimation'
-> & { ref?: React.Ref<SVGPathElement> };
-
-type PieArcInterpolatedProps = Pick<PieArcAnimatedProps, 'startAngle' | 'endAngle'>;
-
-function pieArcPropsInterpolator(from: PieArcInterpolatedProps, to: PieArcInterpolatedProps) {
-  const interpolateStartAngle = interpolateNumber(from.startAngle, to.startAngle);
-  const interpolateEndAngle = interpolateNumber(from.endAngle, to.endAngle);
-
-  return (t: number) => {
-    return {
-      startAngle: interpolateStartAngle(t),
-      endAngle: interpolateEndAngle(t),
-    };
-  };
-}
-
-/** Animates a pie slice using a `path` element.
- * The props object also accepts a `ref` which will be merged with the ref returned from this hook. This means you can
- * pass the ref returned by this hook to the `path` element and the `ref` provided as argument will also be called. */
-export function useAnimatePieArc(props: PieArcAnimatedProps): {
-  ref: React.Ref<SVGPathElement>;
-  d: string;
-} {
-  const ref = useAnimate(
-    { startAngle: props.startAngle, endAngle: props.endAngle },
-    {
-      createInterpolator: pieArcPropsInterpolator,
-      applyProps(element, animatedProps) {
-        element.setAttribute(
-          'd',
-          d3Arc()
-            .cornerRadius(props.cornerRadius)({
-              padAngle: props.paddingAngle,
-              innerRadius: props.innerRadius,
-              outerRadius: props.outerRadius,
-              startAngle: animatedProps.startAngle,
-              endAngle: animatedProps.endAngle,
-            })!
-            .toString(),
-        );
-      },
-      initialProps: {
-        startAngle: (props.startAngle + props.endAngle) / 2,
-        endAngle: (props.startAngle + props.endAngle) / 2,
-      },
-      skip: props.skipAnimation,
-    },
-  );
-
-  return {
-    ref: useForkRef(ref, props.ref),
-    d: d3Arc().cornerRadius(props.cornerRadius)({
-      padAngle: props.paddingAngle,
-      innerRadius: props.innerRadius,
-      outerRadius: props.outerRadius,
-      startAngle: props.startAngle,
-      endAngle: props.endAngle,
-    })!,
-  };
-}
 
 const PieArcRoot = styled('path', {
   name: 'MuiPieArc',
