+++ packages/x-charts/src/hooks/animation/index.ts
@@ -1,0 +1,2 @@
+export * from './useAnimatePieArc';
+export * from './useAnimateLine';
