+++ packages/x-charts/src/internals/animation/Transition.tsx
