+++ packages/x-charts/src/internals/animation/animation.ts
